Pipeline now accept list of paths as input
The pipeline now accepts a list of paths or hashes as input. We can
give a list full of hashes, full of paths or a mixture between the two
as input. The function =convertPathToHash(path)= in tools.py is used
to convert paths values (recognized using the function =os.path.isfile(path)=
from the os module. With this method we can use relative and absolute
paths filenames.

ex. We can run =OptimalScienceExtraction= using input:
 - input =
 ["2133e1778dd6f8208763eeeed3a5ae6efd525fabb33a5fdf8809bd77bf89bb2b",
 "626de973a22fe042b8355bfdb868260e1dc13cbbc411f4baaf6730b813e3a26d"]
 - input =
 ["Data/ProcessedData/ExtractedOrders/extracted_flat_orders.fits",
 "626de973a22fe042b8355bfdb868260e1dc13cbbc411f4baaf6730b813e3a26d"]
 - input =  ["Data/ProcessedData/ExtractedOrders/extracted_flat_orders.fits",
 "~/MARVEL/MARVELpipeline/Data/ProcessedData/ExtractedOrders/extracted_science_orders.fits"]