Prepare files to be run with dvc
Added dvc.yaml file and param.yaml file. The param.yaml files is a
configuration file that can be uses to specify the input/output path

Changed and cleaned-up the files so that they can be called from a
dvc.yaml file and read values from a param.yaml file.